Fixed a bunch of warnings from gcc 4.4.5 . Most of these are parenthesis missing to make the precedence of && over || obvious. Ah well, why not...
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@7765 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/fl_boxtype.cxx
@@ -81,19 +81,19 @@ void fl_frame(const char* s, int x, int y, int w, int h) {
   uchar *g = fl_gray_ramp();
   if (h > 0 && w > 0) for (;*s;) {
     // draw top line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_xyline(x, y, x+w-1);
     y++; if (--h <= 0) break;
     // draw left line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_yxline(x, y+h-1, y);
     x++; if (--w <= 0) break;
     // draw bottom line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_xyline(x, y+h-1, x+w-1);
     if (--h <= 0) break;
     // draw right line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_yxline(x+w-1, y+h-1, y);
     if (--w <= 0) break;
   }
@@ -115,19 +115,19 @@ void fl_frame2(const char* s, int x, int y, int w, int h) {
   uchar *g = fl_gray_ramp();
   if (h > 0 && w > 0) for (;*s;) {
     // draw bottom line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_xyline(x, y+h-1, x+w-1);
     if (--h <= 0) break;
     // draw right line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_yxline(x+w-1, y+h-1, y);
     if (--w <= 0) break;
     // draw top line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_xyline(x, y, x+w-1);
     y++; if (--h <= 0) break;
     // draw left line:
-    fl_color(g[*s++]);
+    fl_color(g[(int)*s++]);
     fl_yxline(x, y+h-1, y);
     x++; if (--w <= 0) break;
   }
